requirements.txt
# Core framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Async HTTP
httpx==0.28.1

# Redis (async support via redis.asyncio)
redis==5.2.1

# Anthropic SDK
anthropic==0.40.0

# ML / re-ranking
sentence-transformers==3.3.1
rank-bm25==0.2.2

# Pydantic v2
pydantic==2.10.4
pydantic-settings==2.7.0

# Logging
structlog==24.4.0

# Config
python-dotenv==1.0.1

# YAML
pyyaml==6.0.2

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0
httpx==0.28.1

# Linting / formatting
ruff==0.8.6
